tf/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
SQLAlchemy==2.0.29
psycopg2-binary
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
pydantic==1.10.2
numpy==1.24.3
thinc==8.1.12
spacy==3.5.4
PyMuPDF==1.23.7
email-validator
